[REFACT]: DogScrollView 수정
